+++ PyBridge.py
@@ -25,7 +25,7 @@ def Main():
     ## Mac    
     elif Platform == "darwin":
         Mac.Mac()
-
+        
     ## Windows
     elif Platform == "win32":
         Windows.Windows()
